Get Consistent results from command line integration tests
* build script picked up Tests.dll from multiple places, causing it to run the tests atleast 3 times
* Seeder & Cluster now share how they new a client, sets ClusterName and TestMethod http headers (useful to relate calls in fiddler)
* We wait for cluster health yellow before signalling a node has started (not strictly needed but for now seems like a good safeguard)
* Sometimes delete index when index does not exists in seeder was taking forever, guarded this with IndexExists check
* Project's suggest context was sometimes an empty array this caused the the cluster settings update task to stall everything and all the readonly tests to fail because the node could not be started in time.

Getting consistent 9 failing tests now.

+++ src/Tests/Framework/Integration/Bootstrappers/Seeder.cs
@@ -9,13 +9,10 @@ namespace Tests.Framework.Integration
 	{
 		private IElasticClient Client { get; }
 
-		public Seeder(int port)
+		public Seeder(ElasticsearchNode node)
 		{
-			var client = TestClient.GetClient(seederSettings, port);
-			this.Client = client;
+			this.Client = node.Client();
 		}
-
-		private ConnectionSettings seederSettings(ConnectionSettings settings) => settings;
 
 		public void SeedNode()
 		{
@@ -35,9 +32,12 @@ namespace Tests.Framework.Integration
 
 		public void DeleteIndicesAndTemplates()
 		{
-			this.Client.DeleteIndexTemplate("raw_fields");
-			this.Client.DeleteIndex(typeof(Project));
-			this.Client.DeleteIndex(typeof(Developer));
+			if (this.Client.IndexTemplateExists("raw_fields").Exists)
+				this.Client.DeleteIndexTemplate("raw_fields");
+			if (this.Client.IndexExists(Indices<Project>()).Exists)
+				this.Client.DeleteIndex(typeof(Project));
+			if (this.Client.IndexExists(Indices<Developer>()).Exists)
+				this.Client.DeleteIndex(typeof(Developer));
 		}
 
 		public void CreateIndices()
